+++ cocoa-beans-schematic-spigot/src/main/java/net/apartium/cocoabeans/spigot/schematic/prop/format/RedstoneWireConnectionsPropFormat.java
@@ -3,7 +3,7 @@ package net.apartium.cocoabeans.spigot.schematic.prop.format;
 import net.apartium.cocoabeans.schematic.prop.BlockProp;
 import net.apartium.cocoabeans.schematic.prop.format.BlockPropFormat;
 import net.apartium.cocoabeans.spigot.schematic.prop.RedstoneWireConnectionsProp;
-import net.apartium.cocoabeans.utils.FileUtils;
+import net.apartium.cocoabeans.utils.BufferUtils;
 import org.bukkit.block.BlockFace;
 import org.bukkit.block.data.type.RedstoneWire;
 
@@ -31,7 +31,7 @@ public class RedstoneWireConnectionsPropFormat implements BlockPropFormat<Map<Bl
     public BlockProp<Map<BlockFace, RedstoneWire.Connection>> decode(byte[] value) {
         DataInputStream in = new DataInputStream(new ByteArrayInputStream(value));
 
-        return constructor.apply(Collections.unmodifiableMap(FileUtils.readMapOfEnums(
+        return constructor.apply(Collections.unmodifiableMap(BufferUtils.readMapOfEnums(
                 in,
                 BlockFace.class, BlockFace::values,
                 RedstoneWire.Connection.class, RedstoneWire.Connection::values
@@ -72,7 +72,7 @@ public class RedstoneWireConnectionsPropFormat implements BlockPropFormat<Map<Bl
     @Override
     public byte[] encode(BlockProp<?> prop) {
         Map<BlockFace, RedstoneWire.Connection> connections = getConnectionsOrElseThrow(prop);
-        return FileUtils.writeMapOfEnums(connections);
+        return BufferUtils.writeMapOfEnums(connections);
     }
 
 }