+++ cocoa-beans-schematic/src/main/java/net/apartium/cocoabeans/schematic/format/SimpleBlockDataEncoder.java
@@ -13,7 +13,7 @@ import java.io.DataInputStream;
 import java.io.IOException;
 import java.util.*;
 
-import static net.apartium.cocoabeans.utils.FileUtils.*;
+import static net.apartium.cocoabeans.utils.BufferUtils.*;
 
 @ApiStatus.AvailableSince("0.0.46")
 public class SimpleBlockDataEncoder implements BlockDataEncoder {