+++ cocoa-beans-schematic/src/main/java/net/apartium/cocoabeans/schematic/prop/format/EnumPropFormat.java
@@ -7,7 +7,7 @@ import java.io.*;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
-import static net.apartium.cocoabeans.utils.FileUtils.*;
+import static net.apartium.cocoabeans.utils.BufferUtils.*;
 
 @ApiStatus.AvailableSince("0.0.46")
 public abstract class EnumPropFormat<T extends Enum<T>> implements BlockPropFormat<T> {