+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 ## 0.0.46
-- [schematic] Initial schematic API
+- [schematic] Introduce schematic API
 - [scoreboard-spigot] Add support for minecraft 1.21.6
 - [common] Bump Minecraft versions from 1.21.6 to 1.21.10
 - [common] Make MinecraftVersion comparable